new Tooltips
Rightclick on items of a setting now brings up the popup menu.
Tooltips are now comparing: if you hover over the shoes of a set and
there are currently shoes in the group loot that are of the same type
(light, medium, heavy) you will now get a tooltip comparing those two
items. The item of your saved set is marked as »currently equipped«.